parse commandline with Argp
First I considered to use the cxxopts-Library from https://github.com/jarro2783/cxxopts.git,
which is a nice idiomatic C++ implementation with features similar to Boost.ProgramOptions.

However, due to the way we collect and overlay options and config files,
it turned out that all we need is to store recognised options as string into a map;
the Argp function from the GNU C stdlib is enough to get that job done.