prepare a test result report
The report is in markdown and always sent to STDOUT for now;
optionally it can also be written into a file. Its rendering
is based on data collected during execution of all test steps.


Note: with this changeset, a complete pass works now
and the testrunner terminates normally (while obviously
the actual test operations are dummies and not yet implemented)