draft mechanism to wait for a match on the subprocess output
The main thread needs to provide some specific match condition,
and then go into blocking wait until this condition is observed
within the output read from the Yoshimi subprocess.

Typical example would be to wait for the "yoshimi>" prompt,
or for the proper termination of the test invocation.

This requires some coordinated protocol, and can be implemented
with the help of a promise to future communication channel.