do not pass the match result object on blocking wait
Again, this turned out to be a case for YAGNI....
Indeed, it /would/ be nice to pass the matcher,
so matching subexpressions /could/ be extracted,
which /might/ be helpful for /future/ extensions...

But implementing that correct turns out to be rather tricky,
since the RexEx-Match refers to the underlying string by reference,
so we'd also have to move that string to pesistent storage and hand
it over to the other thread.

And moreover, this whole matching feature works reliably only under
very specific conditions, where we know precisely at which point
the subject under test (Yoshimi) is waiting at the moment.
Most notably it is *not* possible to wait for a match in output
that has already passed by a the point when we're activating that match.

So bottom line: if we *ever* need extracting of submatches from the output,
then we're better off implementing that on top of a captured log.

And, please note how much simpler the code becomes...
YAGNI YAGNI YAGNI!!