control logging of Yoshimi output from the main thread
Resolve the preliminary solution and log output only when needed.
Explanation: Since the Progress-Indicator stores log messages into
a STL container, we can not just add lines concurrently from several
threads. Even while this works most of the time on AMD64 machines,
it is not correct and can in theory corrupt the container.

We could play a lot of clever tricks here, or even brute force
mutex protect each logging call -- yet a nifty solution is to
hand over the logger to the other thread precisely when the
main thread is in blocking wait; since installing these
wait conditions is already guarded by a std::atomic<bool>,
we can just "piggyback" the hand-over of the logger.