rework testsuite builder to add global statistic steps
After traversal the testsuite directory and evaluation of all
individual test specifications, a final segment of test steps
is added now to work on overarching timing data extracted from
the preceding tests with timing measurement.

As it turns out, the class Builder can be considered an implementation
detail of the testsuite setup and thus has been moved down into the
implementation code and removed from the header. Moreover, an additional
Mould has been set up, to define those global steps to be appended
at the end of the Testsuite.