establish a tolerance band for each test and detect timing failure
Since we capture a time series of timing measurements for each test case,
the tolerance for "the usual fluctuations" can be based on the variance
of those data points. In addition, we have to account for the fitting error
of the current (local) platform model.

When observing a timing difference beyond this tolerance band, the test fails