Cover basic features of the three primary synth algorithms ("power" branch)
Note: sound generated with a yoshimi executable from "power"-branch,
which is work-in-progress and not yet merged into mainline
Compiled with -O3 -march=native -mtune=native ("BuildForThisCPU")

These baselines seems to be stable now, also on several other machines/compilers.
This was the purpose of the "power"-branch, replace powf() by exp-with-precomputed-logarithm.