Padthread: adapt PAD-Synth to current development changes
With the "padthread" branch, currently we attempt to setup
PADSynth wavetable building as a multithreaded background task.

Since this requires to have a private PRNG for each PADnoteParameters,
obviously the random number sequence is changed for PADnotes.
When back-porting this changes to random number generation alone to the
codebase of Yoshimi v2.1.2.2, only a small change of -80dB remains
(TODO: still to be investigated further)

TODO: capture more test cases to cover PADSynth spectrum generation