Testrunner: Runtime Statistics -- tolerance based solely on past data points
To improve triggering of warnings and errors, calculate the tolerance
band excluding the current data point, i.e. only based on the statistical
fluctuations of past data. Thus, when there is a significant step in
the measured timing data, on first incidence still the old (typically
lower) fluctuations are used, and thus the sudden change is more likely
to trigger an alarm. Subsequent measurements will of course already
see the jump and thus use adjusted averages and a wider tolerance band,
which causes the alarm to "go away" afer some repetitions.

Note: Typically also the trend detector will trigger after a stretch
of changed values, which means there is a second opportunity to notice
the change. After investigating the reason, the user is expected to
either fix the problem or re-capture the baseline, and thus the changed
delta will return close to zero on the long run.