PADSynth timing tests: run explicitly without concurrency
As it turned out, the mere fact that we're handing over data from
a background thread into the Synth thread can cause a slowdown by
factor 3-5. And this happens even while the test itself and the
timing measurement does run only within a single thread.

Remark: buildPAD Muted is the "legacy mode";
here the build is done as in previous Yoshimi versions, ie.
the part is muted and then the wavetable is built synchroneously