Bugfix: ill guided baseline file name
When the testcase-ID is "base", then the filename construction was derailed
and searched just for "baseline.wav" instead of "base-baseline.wav".

As it turned out, the condition to suppress the prefix decoration was too sloppy.
We do want to be able to definie a specific name, but we do not want such a fuzzy contains match.