Breaking Change: adapt runtime.csv storage to improve readability
when working with timing tests and runtime trends,
it is important to grasp the runtime CSV data at a glance.

- timing and moving average are the most important values
- remove clutter by discarding excess precision on indicative values

Unfortunately this change breaks existing historical runtime data
collected within the testsuite directory; so it seems appropriate to
push such change now (wile persumably I am the only active user of the
Testsuite) and to provide a migration script to rewrite existing data.